Re-organize modules for better DX experience
