+++ crates/protocol/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod message;
 pub mod resource;
-pub mod resources;
 
 #[cfg(test)]
 pub mod test_data;
