+++ crates/protocol/src/resource/mod.rs
@@ -1,3 +1,5 @@
+pub mod offering;
+
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use serde_json::to_string;
