+++ crates/protocol/src/resource/offering.rs
