--- crates/protocol/src/resources/mod.rs
@@ -1,1 +1,0 @@
-pub mod offering;
